rptest/treewide: fix ruff F841 errors
This is "unused variable". The fixes are the result of running

```
ruff check . --select F841 --fix --unsafe-fixes
```

and then examining and updating each fix (if necessary) by hand.

Ruff info on the check is at:
https://docs.astral.sh/ruff/rules/unused-variable/

It says the unsafety is due to

> This rule's fix is marked as unsafe because removing an unused
> variable assignment may delete comments that are attached to the
> assignment.

I don't really know what they are referring to: perhaps they mean
comments that affect other analyzers like "ignore" type comments. In
any case, I didn't see any cases of deleting comments that could be
a problem. ISTM there are other obscure reasons this could be unsafe,
such as code which introspects the local variables, using `locals()`
but this is _really_ obscure.

Ruff is cautious when deleting code, often leaving the RHS of the
assignment, only deleting the LHS, e.g., if `r` is unused in:

r = x // y

it would leave

x // y

which is "technically necessary" since x // y might have side effects
such as throwing if `y` is zero, but I deleted all these cases since
we don't care about that. Of course when the RHS did have important
side effects, I left it in (this was very common, e.g., requests to
redpanda like topic creation but where we didn't use the result).

In a few cases I could see obviously where the variable was intended to
be used and fixed it (usually the RHS was copied into a later call, I
replaced it with the variable).

A few other cases:

rptest/redpanda_cloud_tests/high_throughput_test.py:

It seems clear to me that we intended to use the 3 preallocated nodes
for the bencn and worker OMB nodes. This code is not currently running
(test is ignored) so it is safe to "fix" it, which I did.

rptest/tests/schema_registry_test.py:

Perhaps we wanted to check something on this v1_id, including Oren to
check.