tests: Run client-swarm forever in the MTT
The MTT life cycle tests are structured in this way where there is the
generic outer layer that creates the topics and runs the client-swarm
for a given amount of time.

The inner layer is then test specific which implements the test
scenarios like restarting etc.

Previously we would have configured the client-swarms to run for 6
minutes. However, there was no contract or coordination between the
inner test function and the outer layer. Such it could be that the
client-swarms would have stopped already part ways through the actual
inner test.

This is bad because it meant that tests which check the client
throughput would fail as the swarms are gone and also it would remove
the client load during the test.

To fix this just run the client-swarm forever and stop it explicitly
after the inner test is done. This matches what we do in other tests
more closely.

This also avoids client-swarm issues like stuck sub-clients.