storage: replace a few vasserts with throws
Recent changes added throw_if_closed() to address races between Kafka
layer callers and log shutdown, which is particularly important because
partitions may shutdown while Kafka layer calls are in flight.

That said, there are other calls into the log that originate from
outside the storage layer, and it isn't obvious that we never race with
a closing log there either. We hope there aren't, because the other
callers are typically Raft or cluster::partition, and we shut down the
Raft instance and partition before shutting down the log. But this
commit is an attempt to cover our bases just in case, throwing in other
calls that stem from these other non-Kafka layers.

Those are the following, which are all called from Raft.
- append
- truncate
- prefix truncate