rptest: PipeTimeout/TimeoutError in raise_on_crash
We scan the logs for crash indicators when a test fails, but on large
logs this check proceeds at about 500 MB/s, which will hit the 30s
timeout if the log is larger than about 15 GB, which it is in some
scale tests.

Instead, just scan the last 10 MB of the logfile, which seems highly
likely to catch all crashes since these are normally right at the
end of the log (after all, the process is crashing).

Fixes: CORE-14683.