rptest/admin_v2: get protoc gen'd stuff to type
Get this to type check, which involved:

Adjusting the proletariat call to use --create-packages which creates
the __init__.py[i] files needed to make these proper packages and for
typing to work. The actual __init__ files are in the next change.

Disable unused import warnings inside the generated files only, since
we can't prevent the compiler from adding them (they are harmless,
in any case).