http: respect connection close directive
Notably AWS S3 always sends `Connection: close` in response to
_DeleteObjects_ operation. Since we didn't respect it the request we
attempt on the same connection always fails resulting in backoff kicking
in (if retries are allowed) resulting in unnecessary delay.

Fortunately, since our http client always creates a `response_stream`
referencing a client we can hook into the destructor and conditionally
invalidate the client/close our connection end. This way, when we
attempt to reuse it we know that we need to create a connection first.

You can see the problem yourself using rpk self-test with AWS S3.

Separately we should consider setting backoff to 0 for the first retry.