cloud_roles: move auth refresh bg op from cloud_io
This is a more natural package for the functionality and the refactor is
needed to allow pushing down auth concerns into client pool rather than
cloud_io which in turn is required for allowing multiple client pools to
coexist with different auth configs.

A visible change of this refactor is that datalake::credential_manager
related auth refreshes will be logged with the datalake logger rather
than cloud_io. Nice bonus allowing to differentiate the two.