http: Enable TCP keepalive on httpd servers
This enables TCP keepalive on the Admin API, PP, SR and possibly others
that use the `ctx_server`.

We hardcode the defaults that we are currently using on the kafka API
which we have never changed anywhere.

This will prevent issues with leaking connections that go through the
AWS NLBs and get silently dropped (as we have previously already seen on
the Kafka API).