`ct/l1`: allow non-contiguous, extent aligned `replace_objects()`... ...requests in `metastore`
Currently, we assume that all extents provided in a `replace_objects()`
or `compact_objects()` request to the `metastore` are contiguous. We also
assert that the update, as a whole, is aligned to existing `extent`s in
the `metastore`.

This is overly restrictive. Consider the following case, in which we have
extents `[[0,99], [100,199], [200,299], [300,399]]`, and wish to replace
the state with the update `[[100,199], [300,399]]`, perhaps due to a compaction
job that decided to only compact those ranges. Currently, the `metastore`
implementation would reject my request, as the offsets `199` and `300` are
not contiguous.

Update the `metastore` so that the inputs in a `replace_objects()` request
are broken into their contiguous components, and checked individually against
the existing `metastore` state, thus allowing non-contiguous updates like the
above. For an even more complex case, we could attempt to compact the above
extents with an update like `[[0,99], [300,349], [350,399]]`, and the update
would now be accepted, since the contiguous ranges in the update
(`[[0,99], [300,399]]`) are individually still extent aligned.