shadowing/tests: use unique_id suffix for group names
When multiple verifier instances are used in a single test, they share
the same consumer group name. This causes previously committed
offsets to be picked during newer invocations.

For example, if the first invocation reads up to offset 100, the next
invocation will start from offset 100 (or later), depending on what was
committed. This leads to flaky behavior in `test_topic_delete`.