`ct/l1`: tighten invariants around `log_compaction_meta` state
We could accidentally queue up the same `log` within the `scheduler`'s
`log_compaction_queue` multiple times.

When re-sampling a queued (but not inflight) `log`'s compaction info from
the `metastore`, we would again push the `log` to the `compaction_queue`.

We still want to resample and re-assign the queued `log`'s updated compaction
info, but we don't want multiple of the same `log` in the queue; this is
bad in principal, and also can lead to concurrency issues with reseting
the `log`'s info after completing a compaction (can lead to crashes/UB,
since the `scheduling_policy` of the `log_compaction_queue` requires this
optional has a value for ordering).

Add a new `log_state` field to `log_compaction_meta`, and use/assert on
this condition in several places to confirm we do not have any concurrency
issues with log management.