cloud_storage_clients: split base transport config from client configuration
I want to make client configuration copyable across shards to make it
easier to work with it. To do that I need to get rid of shared pointers.

As a first step, extract transport configuration (and tls certificate
from it). The potential downside is that we delay slightly error
propagation from building tls credentials. That's acceptable as we start
client pool shortly after starting configs.

I'll extract the client probe shared pointer next.