rptest/log_compaction_test: improve tx removal upgrade tests
One of the scenarios needing may_have_transaction_data_or_fence_batches
segment index flag is the following:
- start on older version
- segment 1 to contain tx fence batch
- segment 2 to contain tx data
- compact them all
- upgrade
- produce commit/abort batches to segment 3.
This proves the need for the flag to clear xact data from segment 2.

Another test added is a "2-segment scenario"
- start on older version
- segment 1 to contain tx fence batch
- segment 2 to contain tx commit batch
- cleanly compact segment 1
- upgrade the cluster
- replicate segment 1 to the new replica: it will contain fence batch
- remove tx batches on the leader
- let segment 2 recover on the replica: it won't contain commit batch.
This proves the need for separate coordination of offsets for tx data.

In the test both scenarios use more segments, as producing or recovering
a specific number of segments is not straightforward, we only can
control these by estimating data sizes and transfer rates.