ct/l0: Introduce accurate per-stage write pipeline accounting
Previously, the write pipeline maintained a _current_sz member that
counted the size of all write requests in the pipeline. It was
incremented just before the write was added to the pipeline and
decremented when request finished and its owning write_and_debounce call
returned. The big problem with _current_sz is that it was being used for
the data threshold, despite counting, e.g., bytes being uploaded,
instead of just bytes waiting to be batched in the scheduler stage.

This change introduces careful accounting of write request bytes by
stage, and uses this accurate accounting in the data threshold. A write
request belongs to the pipeline whose pending request list it is linked
into, and to the stage in its stage member. This means, for example,
that a write request that is being proxied belongs to no pipeline and no
stage, and isn't counted. The proxy request created from it does count.
This differs from _current_sz accounting.

Note that the bytes of write request a shard holds in memory is not the
same thing as the bytes owned by the pipeline, because proxied requests
are still held and accounted for in memory on the source shard, but not
on the target shard. It's the opposite for pipeline stage byte
accounting. It's OK not to count the proxied request because the batch
payload stays in memory on the source shard.