`ct/l1`: add `extent_metadata_reader`
Fetches extents from the `metastore` for the provided `tp` within the range
`[min_offset, max_offset]` inclusively. Iteration (and thereby the order in
which extents are provided) can be forwards or backwards.
For example, for an extent state of `[[0,9],[10,19],[20,29]]` and a
construction like `extent_metadata_reader([0,15,forwards])`, the generator
is expected to yield the extents `[[0,9],[10,19]]`. For a construction like
`extent_metadata_reader([0,15,backwards])`, the generator is expected to
yield `[[10,19],[0,9]]`.

If an error from the `metastore` is received, this error is yielded to the
user. The `extent_metadata_generator` will be left in a valid state if the
user wishes to continue retrying.