ct/l0: Fix time fallback metric
The time fallback metric was counting requests that were proxied, which
left out requests from the target shard. Plus, with the previous change,
proxied requests may be advanced to batching by the data threshold. This
changes the metric so it's counted when the time fallback mechanism
advances requests to the batcher, which matches how the data threshold
counts requests.