lsm: compare file boundaries using user keys in compaction
See the added test, this is required for compaction of tombstones,
otherwise we could incorrectly drop tombstones early, causing values to
be resurface that are deleted.

The fix is to use our new user_key type to enforce we compare by user
keys when comparing file boundaries for determining which files need
to be input to a compaction.