introduce //bazel:sanitizer_mode build flag
Controls sanitizer mode. This replaces a few things like
krb5/sanitizers, has_sanitizers, etc with a single flag that can be
set to "none", "asan" or "all".

The old flag was insufficient as sanitizers are not a boolean state,
since we can now mix and match sanitizers. So settle on 3 modes for now.