ct/reconciler: Make max reconciler object size configurable
There's two good things about this:
1. Increasing the max object size increases the max amount that can be
   reconciled per round, which may be helpful in high throughput
   situations, especially because throughput to S3 tends to very high
   but the first-byte latency is also high.
2. It allows me to use small byte limits in the reconciler test suite
   and therefore the tests run fast.

I also snuck in reducing the number of injected failures when testing
metastore retries. This significantly shortens the test time because
each additional injected failure causes the test to have to wait one
more level of exponential backoff. We are still testing retries in the
case of multiple failures.