ssx: introduce actor
An actor is a single fiber executor, with a few options for the mailbox
behaviors. It is different than the workqueue in that the action is the
same, but you want control over the mailbox size. It is different from
the single fiber executor because there is no preemption when new work
is passed, and there can be a queue of work.