lsm/db: add an actor for file gc
This allows GC to proceed concurrently of the rest of the compaction loop,
as it doesn't need to block new compactions running.

Since it's now going to run concurrently with files currently being
written we need to pass in the ID of the max file committed, so that we
don't delete anything that is starting to be written to the database.
We'll need to make modifications to the version_set to seperately track
the maximum allocated file ID and the maximum committed file ID.